pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "podvoice"
version = "0.1.0"
description = "Local-first CLI to turn Markdown scripts into multi-speaker audio using Coqui XTTS v2."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Podvoice Contributors" }
]
license = { text = "MIT" }
keywords = ["text-to-speech", "tts", "podcast", "audio", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "torch>=2.1.0",
  "torchaudio>=2.1.0",
  "coqui-tts[codec]>=0.27.4",
  "pydub>=0.25.1",
  "typer[all]>=0.9.0",
  "rich>=13.0.0",
]

[project.scripts]
podvoice = "podvoice.cli:app"

[tool.setuptools]
packages = ["podvoice"]
